todoReact/api/requirements/common.txt
Django~=3.1.1
djangorestframework
djangorestframework-jwt
Pillow==7.2.0
django-cors-headers